move dropdown behavior from content part to terrier part
