+++ app/frontend/terrier/parts/terrier-part.ts
@@ -1,10 +1,11 @@
-import {Part, PartTag} from "tuff-core/parts"
+import {Part, PartParent, PartTag} from "tuff-core/parts"
 import {TerrierApp} from "../app"
 import Loading from "../loading"
 import Theme, {IconName} from "../theme"
 import Toasts, {ToastOptions} from "../toasts"
 import {DbErrors} from "../db-client"
 import * as inflection from "inflection"
+import {Dropdown} from "../dropdowns";
 
 /**
  * Base class for ALL parts in a Terrier application.
@@ -156,6 +157,50 @@ export default abstract class TerrierPart<TState> extends Part<TState> {
      */
     successToast(message: string, icon?: IconName) {
         this.showToast(message, {icon, color: 'success'})
+    }
+
+
+    /// Dropdowns
+
+    /**
+     * Shows the given dropdown part on the page.
+     * It's generally better to call `toggleDropdown` instead so that the dropdown will be
+     * hidden upon a subsequent click on the target.
+     * @param constructor a constructor for a dropdown part
+     * @param state the dropdown's state
+     * @param target the target element around which to show the dropdown
+     */
+    makeDropdown<DropdownType extends Dropdown<DropdownStateType>, DropdownStateType extends {}>(
+        constructor: { new(p: PartParent, id: string, state: DropdownStateType): DropdownType; },
+        state: DropdownStateType,
+        target: EventTarget | null) {
+        const dropdown = this.app.addOverlay(constructor, state, 'dropdown')
+        dropdown.parentPart = this
+        if (target && target instanceof HTMLElement) {
+            dropdown.anchor(target)
+            this.app.lastDropdownTarget = target
+        }
+    }
+
+    clearDropdowns() {
+        this.app.clearDropdowns()
+    }
+
+    /**
+     * Calls `makeDropdown` only if there's not a dropdown currently originating from the target.
+     * @param constructor a constructor for a dropdown part
+     * @param state the dropdown's state
+     * @param target the target element around which to show the dropdown
+     */
+    toggleDropdown<DropdownType extends Dropdown<DropdownStateType>, DropdownStateType extends {}>(
+        constructor: { new(p: PartParent, id: string, state: DropdownStateType): DropdownType; },
+        state: DropdownStateType,
+        target: EventTarget | null) {
+        if (target && target instanceof HTMLElement && target == this.app.lastDropdownTarget) {
+            this.clearDropdowns()
+        } else {
+            this.makeDropdown(constructor, state, target)
+        }
     }
 
 }
